+++ app/controllers/api/v1/api_base_controller.rb
@@ -1,0 +1,6 @@
+module Api
+  module V1
+    class ApiBaseController < ApplicationController
+    end
+  end
+end
