+++ app/controllers/api/v1/books_controller.rb
@@ -1,0 +1,40 @@
+class Api::V1::BooksController < Api::V1::ApiBaseController
+  before_action :set_book, only: [:show, :update]
+
+  def index
+    @books = Book.all
+
+    render json: @books
+  end
+
+  def show
+    render json: @book
+  end
+
+  def create
+    @book = Book.new(book_params)
+
+    if @book.save
+      render json: @book, status: :created, location: @book
+    else
+      render json: @book.errors, status: :unprocessable_entity
+    end
+  end
+
+  def update
+    if @book.update(book_params)
+      render json: @book
+    else
+      render json: @book.errors, status: :unprocessable_entity
+    end
+  end
+
+  private
+  def set_book
+    @book = Book.find(params[:id])
+  end
+
+  def book_params
+    params.require(:book).permit(:image_url, :name, :price, :purchage_date)
+  end
+end
