+++ config/routes.rb
@@ -1,4 +1,7 @@
 Rails.application.routes.draw do
-  resources :books
-  # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
+  namespace :api do
+    namespace :v1 do
+      resources :books, except: :destroy
+    end
+  end
 end
